backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
minio==7.2.3
redis==5.0.1
celery[redis]==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
Pillow==10.2.0
pdf2image==1.17.0
python-pptx==0.6.23
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.27.0
email-validator==2.1.0
